+++ commands/textureTools/texture.js
@@ -13,7 +13,10 @@ const { magnify } = require('../../functions/magnify.js');
 const { palette } = require('../../functions/palette.js');
 const { getMeta } = require('../../functions/getMeta.js');
 const { warnUser } = require('../../functions/warnUser.js');
-const { jsonContributionsJava, jsonContributionsBedrock } = require('../../helpers/fileHandler');
+const { timestampConverter } = require ('../../functions/timestampConverter');
+
+const allowed = ['vanilla', '16', '32', '64'];
+const used = ['16', '32', '64'];
 
 module.exports = {
   name: 'texture',
@@ -24,10 +27,6 @@ module.exports = {
   syntax: `${prefix}texture <16/32/64> <texture_name>\n${prefix}texture <16/32/64> <_name>\n${prefix}texture <16/32/64> </folder/>`,
   example: `${prefix}texture 16 dirt`,
   async execute(_client, message, args) {
-
-    const allowed = ['vanilla', 'vanillabedrock', '16', '16j', '16b', '32', '32j', '32b', '64', '64j', '64b', '16x', '16xj', '16xb', '32x', '32xj', '32xb', '64x', '64xj', '64xb'];
-    const java    = ['16', '32', '64'];
-    const bedrock = ['16b', '32b', '64b'];
 
     let results    = []
     const textures = require('../../helpers/firestorm/texture')
@@ -46,14 +45,9 @@ module.exports = {
     else args[1] = String(args[1])
 
     // universal args
-    if (args[0] === 'vanilla') res = '16';
-    if (args[0] === 'vanillabedrock') res = '16b';
-    if (args[0] === '16j' || args[0] === '16xj' || args[0] === '16x') res = '16';
-    if (args[0] === '32j' || args[0] === '32xj' || args[0] === '32x') res = '32';
-    if (args[0] === '64j' || args[0] === '64xj' || args[0] === '64x') res = '64';
-    if (args[0] === '16b' || args[0] === '16xb') res = '16b';
-    if (args[0] === '32b' || args[0] === '32xb') res = '32b';
-    if (args[0] === '64b' || args[0] === '64xb') res = '64b';
+    if (args[0].includes('16') || args[0] === 'vanilla') res = '16'
+    if (args[0].includes('32')) res = '32'
+    if (args[0].includes('64')) res = '64'
 
     // partial texture name (_sword, _axe -> diamond_sword, diamond_axe...)
     if (search.startsWith('_')) {
@@ -80,7 +74,155 @@ module.exports = {
       }])
     }
 
-    console.log(results)
+    if (results.length > 1) getMultipleTexture(message, results, search, res)
+    else getTexture(message, res, results[0])
 
   }
+}
+
+/**
+ * TODO : make this function in it's own file?
+ * Show an embed an let the user choose which texture he wiould to see
+ * @param {DiscordMessage} message discord message
+ * @param {Array} results array of textures
+ * @param {String} search texture name search
+ * @param {String} res texture resolution
+ */
+async function getMultipleTexture(message, results, search, res) {
+  const emoji_num = ['1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣', '6️⃣', '7️⃣', '8️⃣', '9️⃣', '🔟', '🇦', '🇧', '🇨', '🇩', '🇪', '🇫', '🇬', '🇭', '🇮', '🇯']
+
+  var embed = new Discord.MessageEmbed()
+    //.setAuthor('Note: this command isn\'t updated for 21w20a yet')
+    .setTitle(`${results.length} results, react to choose one!`)
+    .setFooter(message.client.user.username, settings.BOT_IMG)
+
+  var description = strings.TEXTURE_SEARCH_DESCRIPTION
+
+  for (let i = 0; i < results.length; i++) {
+    if (i < emoji_num.length) {
+      let uses     = await results[i].uses()
+      let paths    = await uses[0].paths()
+      
+      description += `${emoji_num[i]} — \`[#${results[i].id}]\` ${paths[0].path.replace(search, `**${search}**`).replace(/_/g, '＿')}\n`
+    }
+  }
+  embed.setDescription(description)
+
+  const embedMessage = await message.inlineReply(embed)
+  asyncTools.react(embedMessage, emoji_num.slice(0, results.length))
+
+  const filter_num = (reaction, user) => {
+    return emoji_num.includes(reaction.emoji.name) && user.id === message.author.id
+  }
+
+  embedMessage.awaitReactions(filter_num, { max: 1, time: 60000, errors: ['time'] })
+  .then(async collected => {
+    const reaction = collected.first()
+    if (emoji_num.includes(reaction.emoji.name)) {
+      embedMessage.delete()
+      return getTexture(message, res, results[emoji_num.indexOf(reaction.emoji.name)])
+    }
+  }).catch(async () => {
+    for (let i = 0; i < results.length; i++) {
+      if (i < emoji_num.length) {
+        if (message.channel.type !== 'dm') embedMessage.reactions.cache.get(emoji_num[i]).remove()
+      }
+    }
+  })
+}
+
+/**
+ * TODO: make this function in it's own file?
+ * Show the asked texture
+ * @param {String} res texture resolution
+ * @param {Object} texture
+ */
+async function getTexture(message, res, texture) {
+  var imgURL = undefined;
+
+  const uses  = await texture.uses()
+  const paths = await uses[0].paths()
+  const path  = paths[0].path
+
+  if (res == '16') imgURL = settings.DEFAULT_MC_JAVA_TEXTURE + path;
+  if (res == '32') imgURL = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Compliance-Java-32x/Jappa-1.17/assets/' + path;
+  if (res == '64') imgURL = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Compliance-Java-64x/Jappa-1.17/assets/' + path;
+  
+  if (path.startsWith('textures')) { // hacks to get the right url with bedrock textures
+    if (res == '16') imgURL = settings.DEFAULT_MC_BEDROCK_TEXTURE + path;
+    if (res == '32') imgURL = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Compliance-Bedrock-32x/Jappa-1.16.200/' + path;
+    if (res == '64') imgURL = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Compliance-Bedrock-64x/Jappa-1.16.200/' + path;
+  }
+
+  axios.get(imgURL).then((response) => {
+    getMeta(imgURL).then(async dimension => {
+      const size = dimension.width + '×' + dimension.height;
+
+      var embed = new Discord.MessageEmbed()
+        .setTitle(`[#${texture.id}] ${path}`)
+        .setColor(colors.BLUE)
+        .setURL(imgURL)
+        .setImage(imgURL)
+        .addField('Resolution:', size,true)
+
+      if (res === '16' || res === '16b') embed.setFooter('Vanilla Texture', settings.VANILLA_IMG);
+      if (res === '32' || res === '32b') embed.setFooter('Compliance 32x', settings.C32_IMG)
+      if (res === '64' || res === '64b') embed.setFooter('Compliance 64x', settings.C64_IMG)
+
+      let lastContribution = await texture.lastContribution((res == '32' || res == '64') ? 'c'+res : undefined);
+      let contributors = lastContribution ? lastContribution.contributors.map(contributor => { return `<@!${contributor}>` }) : 'None'
+      let date = lastContribution ? timestampConverter(lastContribution.date) : 'None'
+
+      if (res != '16') {
+        embed.addFields(
+          { name: 'Author(s)', value: contributors, inline: true },
+          { name: 'Added', value: date, inline: true }
+        )
+      }
+
+      embed.addField('Edition', path.startsWith('textures') ? 'Bedrock': 'Java', true)
+
+      const embedMessage = await message.inlineReply(embed);
+      embedMessage.react('🗑️');
+      if (dimension.width <= 128 && dimension.height <= 128) {
+        embedMessage.react('🔎');
+      }
+      embedMessage.react('🌀');
+      embedMessage.react('🎨');
+
+      const filter = (reaction, user) => {
+        return ['🗑️', '🔎', '🌀', '🎨'].includes(reaction.emoji.name) && user.id === message.author.id;
+      };
+
+      embedMessage.awaitReactions(filter, { max: 1, time: 60000, errors: ['time'] })
+      .then(async collected => {
+        const reaction = collected.first()
+        if (reaction.emoji.name === '🗑️') {
+          if (!embedMessage.deleted) embedMessage.delete()
+          if (!message.deleted) message.delete()
+        }
+        if (reaction.emoji.name === '🎨') {
+          return palette(embedMessage, embedMessage.embeds[0].image.url)
+        }
+        if (reaction.emoji.name === '🔎') {
+          return magnify(embedMessage, embedMessage.embeds[0].image.url)
+        }
+        if (reaction.emoji.name === '🌀' && used.includes(res)) {
+          if (!embedMessage.deleted) embedMessage.delete()
+          return getTexture(message, used[(used.indexOf(res) + 1) % used.length], texture)
+        }
+      })
+      .catch(async () => {
+        if (!message.deleted && message.channel.type !== 'dm') await embedMessage.reactions.cache.get('🗑️').remove()
+        if (dimension.width <= 128 && dimension.height <= 128) {
+          if (!message.deleted && message.channel.type !== 'dm') await embedMessage.reactions.cache.get('🔎').remove()
+        }
+        if (!message.deleted && message.channel.type !== 'dm') await embedMessage.reactions.cache.get('🌀').remove()
+        if (!message.deleted && message.channel.type !== 'dm') await embedMessage.reactions.cache.get('🎨').remove()
+      })
+
+    })
+  }).catch((error) => {
+    return warnUser(message, strings.TEXTURE_FAILED_LOADING + '\n' + error)
+  })
 }